requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
requests==2.25.0
zeep==4.0.0
kavenegar==1.1.2
drf-yasg==1.20.0
persian==0.4.0
django-cors-headers==3.5.0
django-nested-inline==0.4.2
openpyxl==3.0.5
python-dateutil==2.8.1
django-wkhtmltopdf==3.3.0
psycopg2-binary==2.8.5
hashids==1.3.1